+++ src/components/stepforms/schemas/index.ts
@@ -11,6 +11,7 @@ import fillnaSchema from './fillna';
 import filterSchema from './filter';
 import formulaSchema from './formula';
 import fromDateSchema from './fromdate';
+import joinSchema from './join';
 import percentageBuildSchema from './percentage';
 import pivotSchema from './pivot';
 import renameBuildSchema from './rename';
@@ -42,6 +43,7 @@ const factories: { [stepname: string]: buildSchemaType } = {
   filter: filterSchema,
   formula: formulaSchema,
   fromdate: fromDateSchema,
+  join: joinSchema,
   lowercase: toLowerSchema,
   percentage: percentageBuildSchema,
   pivot: pivotSchema,